+++ start-all.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+# Same as start.sh
+exec "$(cd "$(dirname "$0")" && pwd)/start.sh" "$@"